jumns-backend/requirements.txt
fastapi>=0.115.0
mangum>=0.19.0
boto3>=1.35.0
strands-agents[gemini]>=0.1.0
strands-agents-tools>=0.1.0
opensearch-py>=2.7.0
requests-aws4auth>=1.3.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
pydantic>=2.10.0
aws-lambda-powertools[all]>=3.0.0
numpy>=1.26.0
